requirements.txt
asgiref==3.2.10
certifi==2020.6.20
Django==3.1.1
django-crontab==0.7.1
mysqlclient==2.0.1
numpy==1.19.2
pandas==1.1.2
pycryptodome==3.9.8
pycryptodomex==3.9.8
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
wincertstore==0.2
xlwt==1.3.0
xlrd==1.2.0
